Split LinktoGeoEnaEga to three classes
